[docker] versioned subfolders should be optional. Resolves #451
Also, tap repositories should be tagged. Resolves #471

+++ core/jreleaser-utils/src/main/java/org/jreleaser/util/FileUtils.java
@@ -221,12 +221,16 @@ public final class FileUtils {
     }
 
     public static void deleteFiles(Path path) throws IOException {
+        deleteFiles(path, false);
+    }
+
+    public static void deleteFiles(Path path, boolean keepRoot) throws IOException {
         if (Files.exists(path)) {
             Files.walk(path)
                 .sorted(Comparator.reverseOrder())
                 .map(Path::toFile)
                 .forEach(File::delete);
-            Files.deleteIfExists(path);
+            if (!keepRoot) Files.deleteIfExists(path);
         }
     }
